+++ src/Neuron/Zettelkasten/Zettel.hs
@@ -33,14 +33,12 @@ instance Ord Zettel where
 instance Show Zettel where
   show Zettel {..} = "Zettel:" <> show zettelID
 
--- TODO: Use generic deriving use field label modifier.
-instance ToJSON Zettel where
-  toJSON Zettel {..} =
-    object
-      [ "id" .= toJSON zettelID,
-        "title" .= zettelTitle,
-        "tags" .= toJSON zettelTags
-      ]
+zettelJson :: KeyValue a => Zettel -> [a]
+zettelJson Zettel {..} =
+  [ "id" .= toJSON zettelID,
+    "title" .= zettelTitle,
+    "tags" .= zettelTags
+  ]
 
 -- | Load a zettel from a file.
 mkZettelFromPath :: FilePath -> Action Zettel