+++ test/Neuron/Zettelkasten/ZettelSpec.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE OverloadedLists #-}
 {-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE QuasiQuotes #-}
+{-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE NoImplicitPrelude #-}
 
 module Neuron.Zettelkasten.ZettelSpec
@@ -25,4 +25,10 @@ spec = do
         dummyContent = either error id $ parsePure "<spec>" "Dummy"
         zettel = Zettel zid "Some title" [Tag "science"] dummyContent
     it "Produces expected json" $ do
-      encode zettel `shouldBe` "{\"id\":\"2011401\",\"title\":\"Some title\",\"tags\":[\"science\"]}"
+      -- "{\"id\":\"2011401\",\"title\":\"Some title\",\"tags\":[\"science\"]}"
+      object (zettelJson zettel)
+        `shouldBe` object
+          [ "id" .= ("2011401" :: Text),
+            "title" .= ("Some title" :: Text),
+            "tags" .= (["science"] :: [Text])
+          ]